+++ thprac_common/src/thprac_gui_input.h
@@ -10,7 +10,7 @@ namespace Gui
 			INGAGME_INPUT_GEN1,
 			INGAGME_INPUT_GEN2,
 	};
-	bool InGameInputInit(ingame_input_gen_t gen, int reg1, int reg2, int reg3 = 0);
+	bool InGameInputInit(ingame_input_gen_t gen, uintptr_t reg1, uintptr_t reg2, uintptr_t reg3 = 0);
 	bool InGameInputGet(int key);
 
 	// Keyboard Input
